Create Neural Network Forward and Backward Propagation Matrix Solver for a Single Example — Regression (Mathematics IA: hkhurram))
This was the code used to calculate all of the updates to the weights and biases, along with the outputs of the neural net in each of the 30 training examples used in supervised learning of the network. The cost function was also evaluated. This code was repeated for each example.